+++ public/images/hero-image.png
