+++ src/components/ConditionalFooter.tsx
